+++ packages/context/src/lib/context-manager/context-permissions/context-permissions-binding.directive.ts
@@ -27,6 +27,17 @@ export class ContextPermissionsBindingDirective implements OnInit, OnDestroy {
 
   @HostListener('addPermission', ['$event'])
   onAddPermission(permission: ContextPermission) {
+    const translate = this.languageService.translate;
+    if (!permission.profil) {
+      const message = translate.instant(
+        'igo.context.contextManager.errors.addPermissionEmpty'
+      );
+      const title = translate.instant(
+        'igo.context.contextManager.errors.addPermissionTitle'
+      );
+      this.messageService.error(message, title);
+      return;
+    }
     const contextId = this.component.context.id;
     this.contextService
       .addPermissionAssociation(
@@ -39,11 +50,15 @@ export class ContextPermissionsBindingDirective implements OnInit, OnDestroy {
           this.component.permissions[permission.typePermission].push(p);
         }
         const profil = permission.profil;
-        const translate = this.languageService.translate;
-        const message = translate.instant('igo.context.permission.dialog.addMsg', {
-          value: profil
-        });
-        const title = translate.instant('igo.context.permission.dialog.addTitle');
+        const message = translate.instant(
+          'igo.context.permission.dialog.addMsg',
+          {
+            value: profil
+          }
+        );
+        const title = translate.instant(
+          'igo.context.permission.dialog.addTitle'
+        );
         this.messageService.success(message, title);
         this.cd.detectChanges();
       });
@@ -64,11 +79,17 @@ export class ContextPermissionsBindingDirective implements OnInit, OnDestroy {
 
         const profil = permission.profil;
         const translate = this.languageService.translate;
-        const message = translate.instant('igo.context.permission.dialog.deleteMsg', {
-          value: profil
-        });
-        const title = translate.instant('igo.context.permission.dialog.deleteTitle');
+        const message = translate.instant(
+          'igo.context.permission.dialog.deleteMsg',
+          {
+            value: profil
+          }
+        );
+        const title = translate.instant(
+          'igo.context.permission.dialog.deleteTitle'
+        );
         this.messageService.success(message, title);
+        this.cd.detectChanges();
       });
   }
 
@@ -108,6 +129,7 @@ export class ContextPermissionsBindingDirective implements OnInit, OnDestroy {
 
   ngOnDestroy() {
     this.editedContext$$.unsubscribe();
+    this.contextService.editedContext$.next(undefined);
   }
 
   private handleEditedContextChange(context: DetailedContext) {