+++ packages/context/src/lib/context-manager/shared/context.service.ts
@@ -27,7 +27,8 @@ import {
   Context,
   DetailedContext,
   ContextMapView,
-  ContextPermission
+  ContextPermission,
+  ContextProfils
 } from './context.interface';
 
 @Injectable({
@@ -46,6 +47,14 @@ export class ContextService {
   // Until the ContextService is completely refactored, this is needed
   // to track the current tools
   private tools: Tool[];
+
+  get defaultContextUri(): string {
+    return this._defaultContextUri || this.options.defaultContextUri;
+  }
+  set defaultContextUri(uri: string) {
+    this._defaultContextUri = uri;
+  }
+  private _defaultContextUri: string;
 
   constructor(
     private http: HttpClient,
@@ -69,18 +78,30 @@ export class ContextService {
     this.readParamsFromRoute();
 
     this.authService.authenticate$.subscribe(authenticated => {
-      const contexts$$ = this.contexts$.subscribe(contexts => {
-        if (contexts$$) {
-          contexts$$.unsubscribe();
+      if (authenticated && this.baseUrl) {
+        this.get().subscribe(contexts => {
           this.handleContextsChange(contexts);
-        }
-      });
-      this.loadContexts();
+        });
+      } else {
+        const contexts$$ = this.contexts$.subscribe(contexts => {
+          if (contexts$$) {
+            contexts$$.unsubscribe();
+            this.handleContextsChange(contexts);
+          }
+        });
+        this.loadContexts();
+      }
     });
   }
 
-  get(): Observable<ContextsList> {
-    const url = this.baseUrl + '/contexts';
+  get(permissions?: string[], hidden?: boolean): Observable<ContextsList> {
+    let url = this.baseUrl + '/contexts';
+    if (permissions && this.authService.authenticated) {
+      url += '?permission=' + permissions.join();
+      if (hidden) {
+        url += '&hidden=true';
+      }
+    }
     return this.http.get<ContextsList>(url);
   }
 
@@ -107,9 +128,27 @@ export class ContextService {
     );
   }
 
+  getProfilByUser(): Observable<ContextProfils[]> {
+    if (this.baseUrl) {
+      const url = this.baseUrl + '/profils?';
+      return this.http.get<ContextProfils[]>(url);
+    }
+    return of([]);
+  }
+
   setDefault(id: string): Observable<any> {
     const url = this.baseUrl + '/contexts/default';
     return this.http.post(url, { defaultContextId: id });
+  }
+
+  hideContext(id: string) {
+    const url = this.baseUrl + '/contexts/' + id + '/hide';
+    return this.http.post(url, {});
+  }
+
+  showContext(id: string) {
+    const url = this.baseUrl + '/contexts/' + id + '/show';
+    return this.http.post(url, {});
   }
 
   delete(id: string): Observable<void> {
@@ -258,10 +297,10 @@ export class ContextService {
     );
   }
 
-  loadContexts() {
+  loadContexts(permissions?: string[], hidden?: boolean) {
     let request;
     if (this.baseUrl) {
-      request = this.get();
+      request = this.get(permissions, hidden);
     } else {
       request = this.getLocalContexts();
     }
@@ -275,17 +314,17 @@ export class ContextService {
       if (!direct && this.baseUrl && this.authService.authenticated) {
         this.getDefault().subscribe(
           (_context: DetailedContext) => {
-            this.options.defaultContextUri = _context.uri;
+            this.defaultContextUri = _context.uri;
             this.addContextToList(_context);
             this.setContext(_context);
           },
           () => {
             this.defaultContextId$.next(undefined);
-            this.loadContext(this.options.defaultContextUri);
+            this.loadContext(this.defaultContextUri);
           }
         );
       } else {
-        this.loadContext(this.options.defaultContextUri);
+        this.loadContext(this.defaultContextUri);
       }
     };
 
@@ -294,7 +333,7 @@ export class ContextService {
         const contextParam = params[this.route.options.contextKey as string];
         let direct = false;
         if (contextParam) {
-          this.options.defaultContextUri = contextParam;
+          this.defaultContextUri = contextParam;
           direct = true;
         }
         loadFct(direct);
@@ -318,6 +357,9 @@ export class ContextService {
       },
       err => {
         contexts$$.unsubscribe();
+        if (uri !== this.options.defaultContextUri) {
+          this.loadContext(this.options.defaultContextUri);
+        }
       }
     );
   }
@@ -352,7 +394,7 @@ export class ContextService {
     this.editedContext$.next(context);
   }
 
-  getContextFromMap(igoMap: IgoMap): DetailedContext {
+  getContextFromMap(igoMap: IgoMap, empty?: boolean): DetailedContext {
     const view = igoMap.ol.getView();
     const proj = view.getProjection().getCode();
     const center: any = new olPoint(view.getCenter()).transform(
@@ -368,22 +410,36 @@ export class ContextService {
         view: {
           center: center.getCoordinates(),
           zoom: view.getZoom(),
-          projection: proj
+          projection: proj,
+          maxZoomOnExtent: igoMap.viewController.maxZoomOnExtent
         }
       },
       layers: [],
       tools: []
     };
 
-    const layers = igoMap.layers$.getValue();
+    let layers = [];
+    if (empty === true) {
+      layers = igoMap.layers$
+        .getValue()
+        .filter(
+          lay =>
+            lay.baseLayer === true ||
+            lay.options.id === 'searchPointerSummaryId'
+        )
+        .sort((a, b) => a.zIndex - b.zIndex);
+    } else {
+      layers = igoMap.layers$.getValue().sort((a, b) => a.zIndex - b.zIndex);
+    }
 
+    let i = 0;
     for (const l of layers) {
       const layer: any = l;
       const opts = {
         id: layer.options.id ? String(layer.options.id) : undefined,
         layerOptions: {
           title: layer.options.title,
-          zIndex: layer.zIndex,
+          zIndex: ++i,
           visible: layer.visible
         },
         sourceOptions: {