+++ packages/geo/src/lib/import-export/shared/export.interface.ts
@@ -4,4 +4,5 @@ export interface ExportOptions {
   format: ExportFormat;
   layer: string;
   name: string;
+  featureInMapExtent: boolean;
 }