+++ packages/integration/src/lib/context/context-editor-tool/context-editor-tool.component.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 
 import { ToolComponent } from '@igo2/common';
+import { ToolState } from '../../tool/tool.state';
 
 @ToolComponent({
   name: 'contextEditor',
@@ -12,4 +13,11 @@ import { ToolComponent } from '@igo2/common';
   selector: 'igo-context-editor-tool',
   templateUrl: './context-editor-tool.component.html'
 })
-export class ContextEditorToolComponent {}
+export class ContextEditorToolComponent {
+
+  constructor(private toolState: ToolState) {}
+
+  submitSuccessed() {
+    this.toolState.toolbox.activatePreviousTool();
+  }
+}