+++ packages/integration/src/lib/filter/spatial-filter-tool/spatial-filter-tool.component.ts
@@ -26,8 +26,10 @@ import { EntityStore, ToolComponent } from '@igo2/common';
 import olFormatGeoJSON from 'ol/format/GeoJSON';
 import { BehaviorSubject } from 'rxjs';
 import { MapState } from '../../map/map.state';
+import { ImportExportState } from './../../import-export/import-export.state';
 import * as olstyle from 'ol/style';
 import { MessageService, LanguageService } from '@igo2/core';
+import { ToolState } from '../../tool/tool.state';
 
 /**
  * Tool to apply spatial filter
@@ -81,7 +83,9 @@ export class SpatialFilterToolComponent {
     private layerService: LayerService,
     private mapState: MapState,
     private messageService: MessageService,
-    private languageService: LanguageService
+    private languageService: LanguageService,
+    private importExportState: ImportExportState,
+    private toolState: ToolState
   ) {}
 
   getOutputType(event: SpatialFilterType) {
@@ -95,6 +99,11 @@ export class SpatialFilterToolComponent {
     if (this.queryType) {
       this.loadFilterList();
     }
+  }
+
+  activateExportTool() {
+    this.importExportState.setSelectedTab(1);
+    this.toolState.toolbox.activateTool('importExport');
   }
 
   private loadFilterList() {
@@ -369,7 +378,7 @@ export class SpatialFilterToolComponent {
    */
   private tryAddLayerToMap(features: Feature[], id) {
     let i = 1;
-    if (features.length > 1) {
+    if (features.length) {
       if (this.map === undefined) {
         return;
       }